+++ adapters/opencv/kleidicv_hal.cpp
@@ -463,6 +463,15 @@ int resize(int src_type, const uchar *src_data, size_t src_step, int src_width,
     return CV_HAL_ERROR_NOT_IMPLEMENTED;
   }
 
+  if (CV_MAT_DEPTH(src_type) == CV_8U && inv_scale_x == 0.5 &&
+      inv_scale_y == 0.5 &&
+      (interpolation == CV_HAL_INTER_LINEAR ||
+       interpolation == CV_HAL_INTER_AREA)) {
+    return convert_error(kleidicv_resize_to_quarter_u8(
+        src_data, src_step, src_width, src_height, dst_data, dst_step,
+        dst_width, dst_height));
+  }
+
   if (interpolation != CV_HAL_INTER_LINEAR) {
     return CV_HAL_ERROR_NOT_IMPLEMENTED;
   }