+++ scripts/ci-opencv.sh
@@ -46,6 +46,7 @@ IMGPROC_TEST_PATTERNS=(
   '*Imgproc_GaussianBlur*'
   '*Imgproc_Sobel*'
   '*Imgproc_Canny*'
+  '*Imgproc_Resize*'
 )
 IMGPROC_TEST_PATTERNS_STR="$(join_strings_with_colon "${IMGPROC_TEST_PATTERNS[*]}")"
 ../../../conformity/opencv_kleidicv/bin/opencv_test_imgproc \